pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "battery-data-standardizer"
version = "0.1.0"
description = "AI-powered automatic standardization of heterogeneous battery data"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "pandas>=1.5",
    "numpy>=1.21",
    "scipy>=1.7",
    "h5py>=3.1",
    "openpyxl>=3.0",
    "click>=8.0",
    "rich>=13.0",
    "PyYAML>=6.0",
    "pydantic>=2.0",
    "vllm>=0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
bds = "bds.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
